check if input is heigher than 0
